# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Kadane Backend API
  description: Kadane Backend API 
  version: 1.0.0
servers:
  - url: https://api.kadane.xyz/v1
    description: Main production server
  - url: http://localhost:3000/v1
    description: Local development server
security: 
  -  Authorization: []
paths:
...
  /accounts:
    get:
      tags:
        - Accounts
      summary: List all accounts with attributes and filters  
      operationId: ListAccountsWithAttributesFiltered 
      parameters:
        - in: query
          name: usernames
          required: false
          schema:
            type: array
            items:
              type: string
          description: Comma-separated list of usernames to include in the response
        - in: query
          name: locations
          required: false
          schema:
            type: array
            items:
              type: string
          description: Comma-separated list of locations to include in the response
        - in: query
          name: sort 
          required: false
          schema:
            type: string
            enum: [level]
            nullable: true
          description: Sort by level
        - in: query
          name: order
          required: false
          schema:
            type: string
            enum: [asc, desc]
          description: Order by ascending or descending
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountsResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
                  
    post:
      tags:
        - Accounts
      summary: Create a new account
      description: Create a new user account with an ID, username, and email.
      operationId: createAccount
      requestBody:
        description: Account data to be created
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAccountRequest'
      responses:
        '201':
          description: Account created successfully
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /accounts/{id}:
    get:
      tags:
        - Accounts
      summary: Get an account by ID
      description: Get an account by ID.
      operationId: getAccountById
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: The ID of the account to get.
      responses:
        '200':
          description: Account response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      tags:
        - Accounts
      summary: Update an account by ID
      description: Update an account by ID.
      operationId: updateAccountById
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: The ID of the account to update.
      requestBody:
        description: Account data to be updated
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountUpdateRequest'
      responses:
        '204':
          description: Account updated successfully
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags:
        - Accounts
      summary: Delete an account by ID
      description: Delete an account by ID.
      operationId: deleteAccountById
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: The ID of the account to delete.
      responses:
        '204':
          description: Account deleted successfully
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /accounts/username/{username}:
    get:
      tags:
        - Accounts
      summary: Get an account by username
      description: Get an account by username.
      operationId: getAccountByUsername
      parameters:
        - in: path
          name: username
          required: true
          schema:
            type: string
          description: The username of the account to get.
      responses:
        '200':
          description: Account response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /accounts/avatar:
    post:
      tags:
        - Accounts
      summary: Upload an avatar image
      description: Upload a base64-encoded avatar image to S3 and store the URL in the accounts table.
      operationId: uploadAvatar
      requestBody:
        description: Base64-encoded avatar image content
        required: true
        content:
          text/plain:
            schema:
              type: string
              example: iVBORw0KGgoAAAANSUhEUgAA...
      responses:
        '201':
          description: Avatar uploaded and URL updated successfully
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes: 
    Authorization:
      type: http
      scheme: bearer
      bearerFormat: jwt 
  schemas:
...
    StatusCode:
      type: string
      enum: [200, 201, 202, 204, 400, 401, 403, 404, 422, 500]
    
    # Responses
    # APIResponse is a generic response that can be used for any response
    # { "data": { any } }
...
    APIErrorData:
      type: object
      properties:
        message:
          type: string
        code:
          type: string
          anyOf:
            - $ref: '#/components/schemas/StatusCode'
    APIError:
      allOf:
        - $ref: '#/components/schemas/APIErrorData'
        - type: object
          properties:
            error:
              type: object
              $ref: '#/components/schemas/APIErrorData'
...
    InternalServerError:
      allOf:
        - $ref: '#/components/schemas/APIError'
        - type: object
          properties:
            code:
              type: string
              enum: [500]

    # Accounts
    Account:
      allOf:
        - type: object
          properties:
            id:
              type: string
            username:
              type: string
            email:
              type: string
              format: email
            avatarUrl:
              type: string
            level:
              type: integer
            createdAt:
              type: string
              format: date-time
            friendStatus:
              type: string
              enum: [none, friend, blocked, request_sent, request_received]
            plan:
              type: string
              enum: [free, plus, pro]
            isAdmin:
              type: boolean
            attributes:
              $ref: '#/components/schemas/AccountAttributes'
          required:
            - id
            - username
            - email
    AccountAttributes:
      description: Account attributes
      type: object
      properties:
        bio:
          type: string
          nullable: true
        contactEmail:
          type: string
          format: email
          nullable: true
        location:
          type: string
          nullable: true
        realName:
          type: string
          nullable: true
        githubUrl:
          type: string
          nullable: true
        linkedinUrl:
          type: string
          nullable: true
        facebookUrl:
          type: string
          nullable: true
        instagramUrl:
          type: string
          nullable: true
        twitterUrl:
          type: string
          nullable: true
        school:
          type: string
          nullable: true
        websiteUrl:
          type: string
          nullable: true
        friendCount:
          type: integer
        blockedCount:
          type: integer
        friendRequestCount:
          type: integer
    AccountUpdateRequest:
      type: object
      properties:
        bio:
          type: string
        contactEmail:
          type: string
          format: email
        location:
          type: string
        realName:
          type: string
        githubUrl:
          type: string
        linkedinUrl:
          type: string
        facebookUrl:
          type: string
        instagramUrl:
          type: string
        twitterUrl:
          type: string
        school:
          type: string
        websiteUrl:
          type: string
    AccountResponse:
      description: Account response
      type: object
      properties:
        data:
          $ref: '#/components/schemas/Account'
    AccountsResponse:
      description: Accounts response
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Account'
...
    CreateAccountRequest:
      description: Create Account Request
      type: object
      properties:
        id:
          type: string
          description: The unique ID for the account.
        username:
          type: string
          description: The username for the account.
        email:
          type: string
          format: email
          description: The email address for the account.
      required:
        - id
        - username
        - email
      example:
        id: "12345"
        username: "john_doe"
        email: "john@example.com"
    # Admin
...
  responses:
...
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/NotFound'
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/InternalServerError'